Created the PlayerController Script and Designed Player movement
Below is the discord comment I wrote to go along with this commit:

I will be pushing the complete player to the main branch now, I anticipate no conflicts yet as I did all of this in a separate scene. Once I get back from work, I will:
- Update the Main Level Scene to use the Player I created.
- Update the Colliders on the Tilemaps to work with the Player.
- Add some folders to the assets folder for things like sprites, prefabs, etc. I will also create the Player prefab.
I will not be adjusting the player jump height or movement speed or anything of the sort as that should be left for those more familiar with the level that is being implemented.