Changed the Level Scene to Include Player Prefab and Fixed Tilemaps Somewhat
Message Copied from Discord:

I have created some of the folders, made the Player Prefab, updated the Tilemaps in the First Level Scene to work better, and also organized the Tilemaps so that details like the flowers are on one Tilemap meanwhile the ground is on its own. The Player does not have full level collision because certain assets don't have a collider attached to them such as the ones in space or the spikes. Once the final level has been mostly solidified we need to add scripts to each interactable object like the spikes and the end of level, that should be separate from the PlayerController script. For now however, the PlayerController script is done and will need minimal changed going forward.